backend/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.2
defusedxml==0.7.1
Deprecated==1.2.14
Django==4.2
django-cors-headers==4.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-social-oauth2==2.1.3
ecdsa==0.18.0
idna==3.4
jwcrypto==1.5.0
mysql-client==0.0.1
oauthlib==3.2.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0

sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0



pip install urllib3==1.26.6 
pip install drf-social-oauth2 asgiref==3.6.0 backports.zoneinfo==0.2.1 Django==3.2 django-cors-headers==3.14.0 djangorestframework==3.14.0 djangorestframework-simplejwt==5.2.2 mysql-client==0.0.1 PyJWT==2.6.0 pytz==2023.3 sqlparse==0.4.4 social-auth-app-django social-auth-core python-decouple
